Update patch set 3
Patch Set 3: Code-Review+2

Better to keep it as it is.  This is a standalone property that we can have a type for.  Anything with the dot after is part of the separate namespace.

My only hesitation on this one is that we basically said we didn't want to make types more strict over time in case someone was using the parts that we'd no longer allow to be set for a property.  Though I guess this one was just a mistake from the beginning given how it's been used, so it's probably fine?

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2