Use a python_binary_host module for insertkeys.py.
Using cc_prebuilt_binary causes the build to use whatever python is
in the caller's path. Use a python_binary_host module for consistent
Python version selection, and update to Python 3 while we're at it.

Most of the Python 3 conversion pieces are the usual str/bytes
disagreements and changed module names. The only complicated issue is
that ParseConfig._out was being GC'd before ParseConfig's finalizer
ran.

The main body is moved into an actual main function to turn its
globals into locals. The finalizer for ParseConfig was being run after
ParseConfig._out was already deleted. This is because ParseConfig._out
was a reference to a global, and globals can be GC'd before
finalizers run when finalizers are run during interpreter shutdown:
https://docs.python.org/3/reference/datamodel.html#object.__del__.

Test: m checkbuild
Bug: None
Change-Id: I974f71639df2ab5bd4fc4316e1228cb1047b398c