Update patch set 1
Patch Set 1:

If this is really a bug with the tools which read /proc/PID/cmdline, shouldn't we be fixing those tools instead?

Processes change their cmdline all the time, and often don't have this capability. The tools need to be correct even in the presence of those processes. Hiding the bug isn't the right solution IMHO.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>