Update patch set 1
Patch Set 1:

Not feasible because:

A) adding a thread that fires every minute for every thread in the system (including kernel threads) that is _also_ locked in memory is not possible. You can not lock one thread without locking all of them.
B) all the threads would need the capability to initiate data collection and kernel panic (they could communicate that to llkd to do the heavy lifting, but that is also wrong minded given constraints).

llkd is effectively [khungtaskd], one of _they_ most powerful threads in the system, but in user space with the configuration and Android requirements to improve the debug and user (mitigation) experience. Being in user space it can be enhanced to do what we are doing here, something that can not be done in the kernel (and be acceptable upstream). It can run on "user" builds, whereas [khungtaskd] can not. It has to be mlocked to monitor the kernel, because it can not be requesting any resources that open itself up to being affected by a kernel live-lock condition. It may also need to be FIFO priority, but we pushed back on that to minimize janking of the system (it takes a few ms to sample the entire system).

Proposal: allow an sepolicy modification to open up /proc/<pid>/stack, but nothing else associated with ptrace? This would require a kernel modification.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>