Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Ok, Plan B:
> 
> - basic llkd services are available in all builds. Checks for processes with no forward scheduling progress in 'D' or 'Z' state, report, kill and/or panic kernel to mitigate and improve health (the alternative is the customer stares at an apparently dead device and long presses the power switch).

It's trivial for an attacker to create a process in a zombie ("Z") state, and we certainly wouldn't want an untrusted program to be able to trick the system into kernel panicing. Zombie processes are just child processes where the parent hasn't explicitly acknowledged the child's death. Zombie processes are very common. 

I don't think there's a similar approach for getting a process into disk ("D") wait, although I wouldn't be surprised if there was. Processes in disk wait are also unkillable, so attempting to kill it doesn't make any sense.

I don't see any problem with exposing process status information to llkd, but I'm really nervous about taking action on the data.

> - Only on userdebug builds provide advanced live lock watchdog services as noted here that observe for kernel stack signatures such as the dependent CLs setup. We may even consider ifdef'ing out the advance code on "user" builds to conserve precious mlockall() memory space.

Seems reasonable.

> - We may consider in later adjustments detecting if there is an active phone call in progress (how? security concerns?), and if so, prevent mitigation until the call disconnects. This is because live-lock generally stops new processes from running, screens from blanking or unblanking, but existing processes (like mlockall'd llkd) continue.

I could see something like this being workable, although the devil is in the details...

> - In two releases, reconsider the possibility of allowing a mature and trusted llkd hardened on dogfood releases to be fully functional on "user" builds.

I'm nervous about exposing a process's stack to a component of Android. Processes, especially user installed programs, have an expectation of confidentiality and integrity of their data, and having access to the stack is a big bypass of those controls. We should talk more in person about this...

> 
> Would that be acceptable alternative plan for now?

Limiting this functionality to userdebug/eng builds seems more reasonable, especially since the purpose of this is to help debug kernel bugs.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>