Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > Ok, Plan B:
> > 
> > - basic llkd services are available in all builds. Checks for processes with no forward scheduling progress in 'D' or 'Z' state, report, kill and/or panic kernel to mitigate and improve health (the alternative is the customer stares at an apparently dead device and long presses the power switch).
> 
> It's trivial for an attacker to create a process in a zombie ("Z") state, and we certainly wouldn't want an untrusted program to be able to trick the system into kernel panicing. Zombie processes are just child processes where the parent hasn't explicitly acknowledged the child's death. Zombie processes are very common.

Read the design document, we first kill the parent, and if we can't after checking a minute later, then proceed to panic and build up a report to diagnose. A DOS attacker would probably have to spawn a child that dies, and somehow figure out how they could go into a 'D' state with no scheduling progress so that they were truly unkillable.

Before, an attacker could spawn threads a plenty until they occupied all process slots with Zombies; this module is being used to deal with such attackers.

> I don't think there's a similar approach for getting a process into disk ("D") wait, although I wouldn't be surprised if there was. Processes in disk wait are also unkillable, so attempting to kill it doesn't make any sense.

'D' processes without scheduling progress are unkillable, it is impossible(tm) for a "user" task to induce such. We send a kill 'to be sure', wait a minute, then consider a panic if that did not work. Disk wait assigned to a user process for 10 minutes is a serious problem, and if that is happening, we need to panic.

Read the design document associated with the bug, we list ways to manage the scope of where we are looking, llkd is not [khungtaskd].

> I don't see any problem with exposing process status information to llkd, but I'm really nervous about taking action on the data.

llkd has been identified as a required health service to A) be able to collect enough triage information to diagnose the live-lock problems, and B) provide a means to stop long press power occurrences that either contribute to bad user experiences or which erases any way to find such information. Even the TRON report will tell us something useful (about unstable products).

> > - Only on userdebug builds provide advanced live lock watchdog services as noted here that observe for kernel stack signatures such as the dependent CLs setup. We may even consider ifdef'ing out the advance code on "user" builds to conserve precious mlockall() memory space.
> 
> Seems reasonable.
> 
> > - We may consider in later adjustments detecting if there is an active phone call in progress (how? security concerns?), and if so, prevent mitigation until the call disconnects. This is because live-lock generally stops new processes from running, screens from blanking or unblanking, but existing processes (like mlockall'd llkd) continue.
> 
> I could see something like this being workable, although the devil is in the details...

Yes it is, and why I signed up for the details ;-}

I suspect we need a 2 hour 1:3 meeting to go over the current state and requirements of llkd. We delivered it with an eye for P, and we are refining it for the next release.

> 
> > - In two releases, reconsider the possibility of allowing a mature and trusted llkd hardened on dogfood releases to be fully functional on "user" builds.
> 
> I'm nervous about exposing a process's stack to a component of Android. Processes, especially user installed programs, have an expectation of confidentiality and integrity of their data, and having access to the stack is a big bypass of those controls. We should talk more in person about this...
> 
> > 
> > Would that be acceptable alternative plan for now?
> 
> Limiting this functionality to userdebug/eng builds seems more reasonable, especially since the purpose of this is to help debug kernel bugs.

Ok, I will switch gears, retest, and update once I am happy with the results.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>