Update patch set 2
Patch Set 2:

+tbao, who has dealt with a lot of these issues from an OTA perspective.

Generally speaking, we only define types in global policy which are intended to be used on all devices, not just one device. So far I've only seen one device which uses this new type. Do you anticipate every device will have one of these block devices with this label?

If this is only intended for 1 device, then declaring the label, and adding the allow rule, should be done only in the device/ tree. No modification to central policy would be needed.

However, if this is something which will be generic to all devices, then it would be better to define this type in core policy, AND move the allow rule from the device specific policy to the core policy. EG, the line:
  allow fastbootd super_block_device:blk_file rw_file_perms;
should not be in the device specific policy, but rather in the global policy.

So, I guess my question for you is: do you anticipate every device will label a "super_block_device"? Is there more than 1 example you can point to where this label will be used?

Patch-set: 2
Reviewer: Gerrit User 1056365 <1056365@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>