Update patch set 6
Patch Set 6:

These are only outgoing IPCs. The main usage of this will be to avoid having to fork a process to communicate with other processes. The instigator of this change is apexd, but it will also be used to send some communications to vold (which are currently time consuming). The relevant details are in the bug, and I'll also attach a link to the doc there.

Patch-set: 6
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>