Update patch set 1
Patch Set 1: Code-Review+1

LGTM.

Brian, for more context: the goal is to remove xt_qtaguid from future devices that ship with Q and kernels >= 4.9. Right now, on devices that launch with P and the 4.9 kernel, xt_qtaguid is already only there for app compatibility; the system doesn't use it any more. We want to delete xt_qtaguid so we can get rid of 3000 lines of out-of-tree kernel code which are a performance, reliability and maintenance burden.

We are taking the approach of restricting access to xt_qtaguid files via selinux instead of simply removing xt_qtaguid on those future devices so that app developers have time to test on existing devices on the Q beta. If we took the approach of simply removing xt_qtaguid on those future devices, then app developers would have no way to notice and fix the problem until those devices are in the field when/after Q ships.

We attempted this last year but it caused at least one major app to crash on startup. Because that happened much late in the release cycle (after P IC date), we partially rolled back the change and replaced it the current selinux policies, which prevent access to the file from any app targeting P or above.

The crashing app has since been fixed, see https://buganizer.corp.google.com/issues/72977484#comment39 . We also added APIs to P to provide alternative ways of accessing the data in this file.

This is our second attempt to remove access to the file from all apps. This time we are making this change well before IC to ensure we have enough time for any applications to be fixed.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1