Remove init writes to /system, tighten neverallows
5c1fe61eaa20d1a47a148e3318229a47e616c834 added a series of new types to
the files on /system. However, because init uses a blacklist to
determine which files it should be able to write, these new types
instantly became writable by init. (In practice, because this partition
is mounted read-only, no harm was possible from granting this)

Remove the ability for init and vendor_init to write to these files on
/system, and update the neverallow rules to prevent regressions.

Test: policy compiles and device boots. No problems.
Change-Id: Ib9f735912cb8c0306ce950977695c86e97582a85