Update patch set 1
Patch Set 1:

hmmm, I don't think we're allowed to make this change. The whole purpose of snapshotting these files is that we've committed to a certain API for treble purposes, and others may be depending on the fact that this is not a string.

trong?

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>