Add whitelist property for updatable driver.
We already have the mechanism to load updatable graphics driver delivered
inside the apk. In order to do conservative updatable graphics driver, we
introduced a whitelist mechanism, where only whitelisted games will use the
updated graphics driver from the apk. To accomodate the need that OEMs may want
to customize their own whitelist, we introduced a new read-only property
ro.gfx.driver.whitelist.0 to let OEM vendors specify the file name of the
whitelist, which lives inside the assets directory of the apk. This property
will be configured when system is built. And the whitelist is extracted during
application launch time when we set up the graphics environment for the
applications.

BUG: 117241379
Test: Build, flash, boot, test with driver apk.

Change-Id: I3db82626bab7f103d38a18982e809c3a7faf1061