Update patch set 1
Patch Set 1:

LGTM in principle, but I expect that lots of apps still depend on this and we'll end up inflicting breakage on dogfooders and inflicting work on teams to triage bugs. Is it feasible to wait a few weeks before merging this?

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>