Update patch set 6
Patch Set 6:

Hi Nick, sorry for the late response. The CL is one of the reference data for the security review b/117477078. Also we made a document in go/android-live-image-security for more detail and the use case. We are stilling waiting for the security process and won't submit. Yes, let's have a discussion next week in MTV.

Patch-set: 6
Reviewer: Gerrit User 1135739 <1135739@85c56323-6fa9-3386-8a01-6480fb634889>