Update patch set 5
Patch Set 5:

> Patch Set 5: Verified+1 Presubmit-Ready+1
> 
> (3 comments)
> 
> Thanks. I think I really didn't understand the role of the cil files. :(

It's for Treble compatibility. It basically is trying to define what label should be used if an older /vendor partition tries accessing this property. Since this property didn't exist in older Android releases, nobody should be referencing it, so it's safe to ignore.

If we had changed labels of the property, and old vendor code might have been accessing it, then your previous change would have been correct.

It's still work-in-progress, but it's our effort to define an SELinux API and deal with labeling compatibility challenges across upgrades.

Patch-set: 5
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>