Update patch set 1
Patch Set 1:

(6 comments)

It looks like a big chunk of this policy was based off the current rules for surfaceflinger. It's unclear from reading the commit message what is the intended behavior of this new process, and how it relates to other, similar processes like surfaceflinger.

Can you help me understand the overall architectural goal for this particular process, and how it relates (replaces?) surfaceflinger? 

Additionally, it may be helpful to view this change in relationship to the other commits, so that the code commits can be examined with the security policy.

Thanks again for answering these questions and helping me get a better understanding of what's going on.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>