Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> So is the intention that only surfaceflinger can talk to the gpu service? Are apps intended to talk to this service too? 

No, this is not the intention of this patch. This service is intended to only access/collect GPU driver capability and report to upper stack. Right now the only customer of this service is when we try to run CTS or/and gaming benchmarks.

> 
> Additionally, in Android, GPU files in /dev are labeled with the gpu_device SELinux label, and are exposed to all applications (https://android.googlesource.com/platform/system/sepolicy/+/9d8f7eac8a2dcc55ad415fceb41deb1dd4286f96/public/app.te#209 line 209). This has been a historic pain point for Android Security, as the GPU drivers are complex kernel code which has a reputation for security bugs. We'd like to remove direct GPU access to apps. Is this new GPU service intended to be the only mechanism that other system components can gain information about the GPU?

I am not sure if this is possible. Apps can write their own rendering code and use GL/Vulkan to do their own rendering. Yes, the GPU service should be the only place where you query GPU driver capability.

Patch-set: 1
Reviewer: Gerrit User 1267775 <1267775@85c56323-6fa9-3386-8a01-6480fb634889>