Update patch set 5
Patch Set 5:

(1 comment)

Resolved offline. Added a comment in .te file.

In summary, the GPU service needs to be able to load the graphics driver in order to query the capability of it. For example right now, it loads the Vulkan driver. The driver is from /vendor/lib[64?]/hw/vulkan.SOC_CHIPSET_NAME.so

Patch-set: 5
Reviewer: Gerrit User 1267775 <1267775@85c56323-6fa9-3386-8a01-6480fb634889>