Update patch set 1
Patch Set 1:

I'm OOO for the rest of this week, so I can't review this in detail.

Without having looked at the policy change itself, my first question is why a new domain is needed at all? If this is app controlled, it feels like this code should be running in the relevant appdomain sandboxes that we already have.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>