Update patch set 7
Patch Set 7:

(1 comment)

> Patch Set 5:
> 
> (2 comments)
> 
> Sorry, I'm missing the background here. Is there a bug or something else which describes why this is needed?
> 
> In particular, how does /dev/cpu_variant differ from /proc/cpuinfo? In situations where cpuinfo is important, we already bind mount /system/lib/<ISA>/cpuinfo on top of /proc/cpuinfo. This is most commonly used for native bridge functionality.

This change is to allow libc select the right routine for different CPUs at runtime. Currently it is the build system who decides which routine to compile in. But we want to build all possible routines in and link the right one at dynamic linking time.

(See https://android-review.googlesource.com/c/platform/bionic/+/822340 for bionic change)

The problem with /proc/cpuinfo or MIDR_EL1 are:
1. In big.little system, it is difficult to automatically decide which cpu to target.
2. Sometimes the choice is kind of arbitrary: even CPU supports some instructions, we may not want to use it.
3. We don't want to maintain a new list of cpu implementer / variant to SOC mapping.

So for now we make the build system write the variant to system property. And make decision based on that information.

Patch-set: 7
Reviewer: Gerrit User 1307957 <1307957@85c56323-6fa9-3386-8a01-6480fb634889>