Create policy for boringssl_self_test
NIAP certification requires that all cryptographic functions
undergo a self-test during startup to demonstrate correct
operation. The boringssl_self_test program performs this check
of BoringSSL.

If the check fails, the new program needs to prevent the device
from booting. To this end, it will cause the device to boot into
the bootloader. The permissions it is given allows it to run
from init and to request a reboot.

Test: Built for walleye. When the BoringSSL self test was forced
      to fail the device rebooted into the bootloader, as
      expected.

Change-Id: I4171b1dd0a5e393252ae5c002171ac51c9cbb3e6