Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > I don't understand why audioserver has to set this by itself and by default. If you need to change the log level for your own device, can you do it in your own init.rc file?
> > > 
> > > Please see discussion in CL744768, 813373 and 820944 as well.
> > > Since "LOG_NDEBUG 0" is defined in AudioPolicyManager in CL744768, default value must be defined for all device(not only our own device), otherwise ALOGV is not silence even in production builds.
> > 
> > I am sorry but I don't agree with aosp/744768 either. I don't own the audio framework, but the desire to keep log even for production builds seems to be a requirement very specific to your devices. It doesn't seem to be a common requirement across the ecosystem. Why can't you do this locally in your fork of Android?
> > 
> > However, since Mikhail +2'ed the audio-side change, I feel less concerned...
> 
> I do not think this is specific for our devices, I believe it's useful for all devices.
> Mikhail, could you give us your comment?

Right, the fact that the change comes from one of the partners doesn't mean it has no use for the rest of the ecosystem, and I appreciate Tomoharu's efforts in going through several design rounds with Colin, Tom, and myself.

Since audio routing policy is affected by the actual configuration and the current state of the device, debugging issues with routing is not easy, so being able to turn on more logging at runtime has a great benefit. Audio policy is only involved in setting up audio connections, thus the overhead of unstripped log statements is negligible.

And the current approach with audio policy manager setting the default log level seems like the easiest that could be currently implemented.

Patch-set: 1
Reviewer: Gerrit User 1042638 <1042638@85c56323-6fa9-3386-8a01-6480fb634889>