Update patch set 1
Patch Set 1:

Technically the allow rule here is more general than the auditallow, as it allows e.g., search, ioctl.  So this could actually remove functionality we didn't see through audit2allow, right?  Is it worth checking for that, or is it unlikely enough that it doesn't matter?

Patch-set: 1
Reviewer: Gerrit User 1208973 <1208973@85c56323-6fa9-3386-8a01-6480fb634889>