Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)

Hi, Steven

First, run as root is not allowed by selinux, we have enforcing selinux in our android build.

So back to your question, we developed some daemons and scripts to collect logs and crashes in android system, these daemons and scripts store, run in vendor partition, from selinux side, if vendor partition service want to call logcat to capture log to filesystem, it need add a domain type in selinux to grant all necessary rights to make sure the call is success and can save logs to persistent location.

This is well managed, only few selinux rule is added with this change and only works in user debug and eng build.

Patch-set: 4
CC: Gerrit User 1398157 <1398157@85c56323-6fa9-3386-8a01-6480fb634889>