Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 2:
> 
> (3 comments)
> 
> In other conversations that I've had with zhaojiac, I was told that we don't want to have the OTA go directly to the system, but rather, through GMS core. This was because GMS core did a bunch of other work like checking battery level, providing UI, etc.
> 
> I'm really getting confused by our numerous OTA mechanisms here. We have:
> 
> 1) adb sideload
> 2) GMSCore Driven updates via priv-app APIs
> 3) updates via fastboot
> 4) This update mechanism
> 5) An update mechanism gkaiser is working on...

This feature is about adding a public API for OTAs to be installed in a constrained environment (large number of corp-owned device fleet where distributing OTAs over the internet is either impossible due to lack of connectivity or will cause significant network congestion). IIUC none of the other methods are public APIs and can be invoked by a designated app (Device Owner in our case). As this is an API, we explicitly don't want it to trigger UI. Battery level is also checked as part of the API contract and will fail the installation if it's too low.

Patch-set: 2
Reviewer: Gerrit User 1059499 <1059499@85c56323-6fa9-3386-8a01-6480fb634889>