Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (5 comments)
> > 
> > Providing more context:
> > CL: ag/5344064
> > Design doc: go/q-local-ota-design
> > 
> > Our feature is about giving admins support to install update from local file instead of automated OTAs. They have update file that they could push locally (to prevent disruption of network or the devices itself) and then call this API in time that it's suitable for them, taking it to account that this is use-case for environment with large fleets of dedicated devices.
> 
> How does interact with updates performed directly from the manufacturer, eg GmsCore in Google's case? Does this new API provide a mechanism to suppress GmsCore provided updates? If so, then the existence of this API would be a denial of service attack stopping users from getting updates. If not, then the API doesn't actually provide the organization any control, since GmsCore can decide to do updates on it's own without involvement from the organization.
> 
> Or is this feature only intended for non-certified devices where images are created and distributed by the organization? If so, then why not just use the same priv-app APIs that we already provide?

This feature is for "Managed system updates" (https://developer.android.com/work/dpc/system-updates), which takes a priority over manufacturers' system updaters (e.g. GmsCore). It basically blocks the updates from the system updaters (this is already happening, with the DevicePolicyManager API added in P and prior), and now it tries to invoke an install with the new API, as an updater.

The reason for not using the same priv-app APIs: DevicePolicyManager wants _public_ APIs; and it wants to abstract away the platform-specific bits.

Patch-set: 2
Reviewer: Gerrit User 1056365 <1056365@85c56323-6fa9-3386-8a01-6480fb634889>