Update patch set 5
Patch Set 5:

Note: Other parts of the topic conflict with internal code. If that continues I'll probably decouple this one from the rest of the topic so I have less to cherry-pick. I think the conflict is due to http://aosp/829882 not being in aosp/master yet, so once that lands and I rebase it might land cleanly.

Patch-set: 5
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>