Update patch set 1
Patch Set 1: Code-Review+2

I asked this in the bug, but I'll also ask this here: this feature only looks needed if we allow toggling between randomizing and not-randomizing addresses. Randomization has lots of security benefits so I'm not sure why we'd support not randomizing it. Plus this is a unique identifier, so I'd prefer not to expose it to anyone.

If we really need it, then I'm fine with it. It looks like this is an unprivileged operation on most non-Android systems, so it seems mostly benign, especially because it's only applicable to this domain.

So I'm fine with it *if* we really need to support non-randomized addresses, but I think we can do better.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2